softlayer: add domain actions to other methods
