+++ go/src/koding/kites/kloud/provider/softlayer/stop.go
@@ -39,29 +39,3 @@ func (m *Machine) Stop(ctx context.Context) error {
 
 	return m.MarkAsStoppedWithReason("Machine is stopped")
 }
-
-func (m *Machine) deleteDomains() error {
-	m.push("Initializing domain instance", 65, machinestate.Stopping)
-	if err := m.Session.DNSClient.Validate(m.Domain, m.Username); err != nil {
-		return err
-	}
-
-	m.push("Changing domain to sleeping mode", 85, machinestate.Stopping)
-	if err := m.Session.DNSClient.Delete(m.Domain); err != nil {
-		m.Log.Warning("couldn't delete domain %s", err)
-	}
-
-	// also get all domain aliases that belongs to this machine and unset
-	domains, err := m.Session.DNSStorage.GetByMachine(m.ObjectId.Hex())
-	if err != nil {
-		m.Log.Error("fetching domains for unseting err: %s", err.Error())
-	}
-
-	for _, domain := range domains {
-		if err := m.Session.DNSClient.Delete(domain.Name); err != nil {
-			m.Log.Warning("couldn't delete domain %s", err)
-		}
-	}
-
-	return nil
-}
